small refactoring and add method parsing
