+++ Sources/GASTBuilder/Tree/Services/OperationNode.swift
@@ -1,0 +1,15 @@
+//
+//  OperationNode.swift
+//  
+//
+//  Created by Александр Кравченков on 14.12.2020.
+//
+
+import Foundation
+
+public struct OperationNode {
+    public let method: String
+    public let description: String?
+    public let summary: String?
+    public let parameters: [ParameterNode]
+}
