+++ Sources/GASTBuilder/Tree/Services/PathNode.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  
+//
+//  Created by Александр Кравченков on 14.12.2020.
+//
+
+import Foundation
+
+public struct PathNode {
+    public var path: String
+    public var operations: [OperationNode]
+}
